+++ examples/matrix_task_processor/matrix.h
@@ -16,7 +16,7 @@ int ** AllocMatrix(int r, int c);
 void FreeMatrix(int ** a, int r, int c);
 void GenMatrix(int ** matrix, const int height, const int width);
 void GenMatrixType(int ** matrix, const int height, const int width, int type);
-int AvgElement(int ** matrix, const int height, const int width);
+double AvgElement(int ** matrix, const int height, const int width);
 int SumMatrix(int ** matrix, const int height, const int width);
 void DisplayMatrix(int ** matrix, const int height, const int width, FILE *stream);
